+++ magic.c
@@ -32,7 +32,7 @@ int dice (int number, int size);
 
 /* Offensive Spells */
 
-void spell_magic_missile (byte level, struct char_data *ch,
+void spell_magic_missile (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -54,7 +54,7 @@ void spell_magic_missile (byte level, struct char_data *ch,
 
 
 
-void spell_chill_touch (byte level, struct char_data *ch,
+void spell_chill_touch (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -85,7 +85,7 @@ void spell_chill_touch (byte level, struct char_data *ch,
 
 
 
-void spell_burning_hands (byte level, struct char_data *ch,
+void spell_burning_hands (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -109,7 +109,7 @@ void spell_burning_hands (byte level, struct char_data *ch,
 
 
 
-void spell_shocking_grasp (byte level, struct char_data *ch,
+void spell_shocking_grasp (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -132,7 +132,7 @@ void spell_shocking_grasp (byte level, struct char_data *ch,
 
 
 
-void spell_lightning_bolt (byte level, struct char_data *ch,
+void spell_lightning_bolt (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -155,7 +155,7 @@ void spell_lightning_bolt (byte level, struct char_data *ch,
 
 
 
-void spell_colour_spray (byte level, struct char_data *ch,
+void spell_colour_spray (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -178,7 +178,7 @@ void spell_colour_spray (byte level, struct char_data *ch,
 
 
 /* Drain XP, MANA, HP - caster gains HP and MANA */
-void spell_energy_drain (byte level, struct char_data *ch,
+void spell_energy_drain (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam, xp, mana;
@@ -217,7 +217,7 @@ void spell_energy_drain (byte level, struct char_data *ch,
 
 
 
-void spell_fireball (byte level, struct char_data *ch,
+void spell_fireball (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -240,7 +240,7 @@ void spell_fireball (byte level, struct char_data *ch,
 
 
 
-void spell_earthquake (byte level, struct char_data *ch,
+void spell_earthquake (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -269,7 +269,7 @@ void spell_earthquake (byte level, struct char_data *ch,
 
 
 
-void spell_dispel_evil (byte level, struct char_data *ch,
+void spell_dispel_evil (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -298,7 +298,7 @@ void spell_dispel_evil (byte level, struct char_data *ch,
 
 
 
-void spell_call_lightning (byte level, struct char_data *ch,
+void spell_call_lightning (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -321,7 +321,7 @@ void spell_call_lightning (byte level, struct char_data *ch,
 
 
 
-void spell_harm (byte level, struct char_data *ch,
+void spell_harm (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -345,7 +345,7 @@ void spell_harm (byte level, struct char_data *ch,
 
 /* spells2.c - Not directly offensive spells */
 
-void spell_armor (byte level, struct char_data *ch,
+void spell_armor (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -367,7 +367,7 @@ void spell_armor (byte level, struct char_data *ch,
 
 
 
-void spell_teleport (byte level, struct char_data *ch,
+void spell_teleport (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int to_room;
@@ -394,7 +394,7 @@ void spell_teleport (byte level, struct char_data *ch,
 
 
 
-void spell_bless (byte level, struct char_data *ch,
+void spell_bless (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -430,7 +430,7 @@ void spell_bless (byte level, struct char_data *ch,
 
 
 
-void spell_blindness (byte level, struct char_data *ch,
+void spell_blindness (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -461,7 +461,7 @@ void spell_blindness (byte level, struct char_data *ch,
 
 
 
-void spell_clone (byte level, struct char_data *ch,
+void spell_clone (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
 
@@ -479,7 +479,7 @@ void spell_clone (byte level, struct char_data *ch,
 
 
 
-void spell_control_weather (byte level, struct char_data *ch,
+void spell_control_weather (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   /* Control Weather is not possible here!!! */
@@ -488,7 +488,7 @@ void spell_control_weather (byte level, struct char_data *ch,
 
 
 
-void spell_create_food (byte level, struct char_data *ch,
+void spell_create_food (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct obj_data *tmp_obj;
@@ -524,7 +524,7 @@ void spell_create_food (byte level, struct char_data *ch,
 
 
 
-void spell_create_water (byte level, struct char_data *ch,
+void spell_create_water (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int water;
@@ -566,7 +566,7 @@ void spell_create_water (byte level, struct char_data *ch,
 
 
 
-void spell_cure_blind (byte level, struct char_data *ch,
+void spell_cure_blind (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   assert (victim);
@@ -581,7 +581,7 @@ void spell_cure_blind (byte level, struct char_data *ch,
 
 
 
-void spell_cure_critic (byte level, struct char_data *ch,
+void spell_cure_critic (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int healpoints;
@@ -603,7 +603,7 @@ void spell_cure_critic (byte level, struct char_data *ch,
 
 
 
-void spell_cure_light (byte level, struct char_data *ch,
+void spell_cure_light (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int healpoints;
@@ -625,7 +625,7 @@ void spell_cure_light (byte level, struct char_data *ch,
 
 
 
-void spell_curse (byte level, struct char_data *ch,
+void spell_curse (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -664,7 +664,7 @@ void spell_curse (byte level, struct char_data *ch,
 
 
 
-void spell_detect_evil (byte level, struct char_data *ch,
+void spell_detect_evil (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -688,7 +688,7 @@ void spell_detect_evil (byte level, struct char_data *ch,
 
 
 
-void spell_detect_invisibility (byte level, struct char_data *ch,
+void spell_detect_invisibility (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -712,7 +712,7 @@ void spell_detect_invisibility (byte level, struct char_data *ch,
 
 
 
-void spell_detect_magic (byte level, struct char_data *ch,
+void spell_detect_magic (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -735,7 +735,7 @@ void spell_detect_magic (byte level, struct char_data *ch,
 
 
 
-void spell_detect_poison (byte level, struct char_data *ch,
+void spell_detect_poison (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   assert (ch && (victim || obj));
@@ -764,7 +764,7 @@ void spell_detect_poison (byte level, struct char_data *ch,
 
 
 
-void spell_enchant_weapon (byte level, struct char_data *ch,
+void spell_enchant_weapon (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int i;
@@ -801,7 +801,7 @@ void spell_enchant_weapon (byte level, struct char_data *ch,
 
 
 
-void spell_heal (byte level, struct char_data *ch,
+void spell_heal (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   assert (victim);
@@ -819,7 +819,7 @@ void spell_heal (byte level, struct char_data *ch,
 }
 
 
-void spell_invisibility (byte level, struct char_data *ch,
+void spell_invisibility (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -849,7 +849,7 @@ void spell_invisibility (byte level, struct char_data *ch,
 }
 
 
-void spell_locate_object (byte level, struct char_data *ch,
+void spell_locate_object (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct obj_data *i;
@@ -889,7 +889,7 @@ void spell_locate_object (byte level, struct char_data *ch,
 }
 
 
-void spell_poison (byte level, struct char_data *ch,
+void spell_poison (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -918,7 +918,7 @@ void spell_poison (byte level, struct char_data *ch,
 }
 
 
-void spell_protection_from_evil (byte level, struct char_data *ch,
+void spell_protection_from_evil (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -937,7 +937,7 @@ void spell_protection_from_evil (byte level, struct char_data *ch,
 }
 
 
-void spell_remove_curse (byte level, struct char_data *ch,
+void spell_remove_curse (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -963,7 +963,7 @@ void spell_remove_curse (byte level, struct char_data *ch,
 }
 
 
-void spell_remove_poison (byte level, struct char_data *ch,
+void spell_remove_poison (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
 
@@ -987,7 +987,7 @@ void spell_remove_poison (byte level, struct char_data *ch,
 
 
 
-void spell_sanctuary (byte level, struct char_data *ch,
+void spell_sanctuary (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -1008,7 +1008,7 @@ void spell_sanctuary (byte level, struct char_data *ch,
 
 
 
-void spell_sleep (byte level, struct char_data *ch,
+void spell_sleep (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -1035,7 +1035,7 @@ void spell_sleep (byte level, struct char_data *ch,
 
 
 
-void spell_strength (byte level, struct char_data *ch,
+void spell_strength (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -1056,7 +1056,7 @@ void spell_strength (byte level, struct char_data *ch,
 
 
 
-void spell_ventriloquate (byte level, struct char_data *ch,
+void spell_ventriloquate (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   /* Not possible!! No argument! */
@@ -1064,7 +1064,7 @@ void spell_ventriloquate (byte level, struct char_data *ch,
 
 
 
-void spell_word_of_recall (byte level, struct char_data *ch,
+void spell_word_of_recall (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   extern int top_of_world;
@@ -1101,7 +1101,7 @@ void spell_word_of_recall (byte level, struct char_data *ch,
 }
 
 
-void spell_summon (byte level, struct char_data *ch,
+void spell_summon (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   sh_int target;
@@ -1131,7 +1131,7 @@ void spell_summon (byte level, struct char_data *ch,
 }
 
 
-void spell_charm_person (byte level, struct char_data *ch,
+void spell_charm_person (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -1184,7 +1184,7 @@ void spell_charm_person (byte level, struct char_data *ch,
 
 
 
-void spell_sense_life (byte level, struct char_data *ch,
+void spell_sense_life (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   struct affected_type af;
@@ -1208,7 +1208,7 @@ void spell_sense_life (byte level, struct char_data *ch,
  *                     Not cast-able spells                                  *
  * ************************************************************************* */
 
-void spell_identify (byte level, struct char_data *ch,
+void spell_identify (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   char buf[256], buf2[256];
@@ -1357,7 +1357,7 @@ void spell_identify (byte level, struct char_data *ch,
  *                     NPC spells..                                          *
  * ************************************************************************* */
 
-void spell_fire_breath (byte level, struct char_data *ch,
+void spell_fire_breath (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -1397,7 +1397,7 @@ void spell_fire_breath (byte level, struct char_data *ch,
 }
 
 
-void spell_frost_breath (byte level, struct char_data *ch,
+void spell_frost_breath (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -1436,7 +1436,7 @@ void spell_frost_breath (byte level, struct char_data *ch,
 }
 
 
-void spell_acid_breath (byte level, struct char_data *ch,
+void spell_acid_breath (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -1481,7 +1481,7 @@ void spell_acid_breath (byte level, struct char_data *ch,
 }
 
 
-void spell_gas_breath (byte level, struct char_data *ch,
+void spell_gas_breath (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
@@ -1505,7 +1505,7 @@ void spell_gas_breath (byte level, struct char_data *ch,
 }
 
 
-void spell_lightning_breath (byte level, struct char_data *ch,
+void spell_lightning_breath (int level, struct char_data *ch,
   struct char_data *victim, struct obj_data *obj)
 {
   int dam;
